requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
idna==3.3
igramscraper==0.3.5
Pillow==9.0.0
python-slugify==3.0.2
requests==2.27.1
text-unidecode==1.2
urllib3==1.26.8
